Move mesh generation to Castaway.Rendering
